+++ apps/playground/src/app/components/components.module.ts
@@ -8,9 +8,23 @@ const routes: Routes = [
       import('./datetime/datetime.module').then((m) => m.DatetimeModule),
   },
   {
+    path: 'flyout',
+    loadChildren: () =>
+      import('./flyout/flyout.module').then((m) => m.FlyoutModule),
+  },
+  {
+    path: 'layout',
+    loadChildren: () =>
+      import('./layout/layout.module').then((m) => m.LayoutModule),
+  },
+  {
     path: 'lookup',
     loadChildren: () =>
       import('./lookup/lookup.module').then((m) => m.LookupModule),
+  },
+  {
+    path: 'tabs',
+    loadChildren: () => import('./tabs/tabs.module').then((m) => m.TabsModule),
   },
 ];