+++ apps/playground/src/app/components/tabs/vertical-tabset/vertical-tabset-routing.module.ts
@@ -1,0 +1,16 @@
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+
+import { VerticalTabsetComponent } from './vertical-tabset.component';
+
+@NgModule({
+  imports: [
+    RouterModule.forChild([
+      {
+        path: '',
+        component: VerticalTabsetComponent,
+      },
+    ]),
+  ],
+})
+export class VerticalTabsetRoutingModule {}